README: remove the copyright notice
The copyright notice would need updating, because there are also others
than Lennart and Colin who own copyrights to paprefs, and the years are
out of date. Having to continuously update the copyright years seems
just unnecessary churn to me. People can browse the git history if
they're interested in the copyright details.