mainwindow: change "[Set as] fallback" to "default"
The "Set as fallback" title is a frequent source of confusion, many
people look at it and do not realize that its actual functional is
"setting as default". Presumably, when this title first appeared it
was presumed that apps should chose a device themselves, and then if
they don't there's a "fallback". But in reality most apps do not chose
any device explicitly, and it's expected behavior to make use of this
"fallback" device. So it is more correct to call a "default" device.