desktop: Match the name in the appstream file
With the addition of the appstream file, the app is showing up in Flathub
and software apps (like GNOME Software) as "Volume Control", but on the
desktop (using the Desktop file), it's showing up differently as
"PulseAudio Volume Control", which is confusing.

Like commit be02260941 mentions, there are some guidelines for naming apps
which prefer the shorter "Volume Control" more than "PulseAudio
Volume Control". It also helps for example GNOME users who get an
ellipsized name in GNOME Software and the GNOME desktop.

To make sure people still get a hit on "PulseAudio" when searching for
the app, let's add that specifically to the tags. That should prevent
search from breaking in a lot of cases.

Part-of: <https://gitlab.freedesktop.org/pulseaudio/pavucontrol/-/merge_requests/97>